rpk/cli: Add 'rpk topic consume', depr. 'rpk api consume'
